Adding LFS support for models. Adding first-class batching support. Updating sample to allow for real batching. Removing implementation details from crate naming for reference implementation.

+++ nv-view/src/transform.rs
@@ -31,7 +31,7 @@ impl GlobalTransformEstimate {
     pub fn displacement_magnitude(&self) -> f32 {
         let tx = self.transform.m[2];
         let ty = self.transform.m[5];
-        ((tx * tx + ty * ty) as f64).sqrt() as f32
+        (tx * tx + ty * ty).sqrt() as f32
     }
 }
